zephyr: Fix rcutils_get_env() again
The commit dacf88e1f4ad9 changed its behaviour by setting env_value
but it only changes the pointer value when it's NULL.  This is wrong.
Instead, we must check that we have given a valid pointer to a
pointer, then we can update it.

Also the return value of the function is "error" string. NULL means
success.  Even though we do not support environment variables in
Zephyr, we must pretend the function call succeeds and an environment
variable the callers are looking for is not set by setting env_value
to "". Otherwise, some functions in rclc or any other components treat
it an error.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>